+++ pass-manager-tkinter-encrypted/data/readme.txt
@@ -1,0 +1,1 @@
+This directory will store all your data.
