+++ pass-manager-tkinter-encrypted/launcher.py
@@ -1,0 +1,3 @@
+from scripts import ui
+
+ui_obj = ui.UI()
