--- pass-manager-tkinter/data/readme.txt
@@ -1,1 +1,0 @@
-This directory will store all your data.
