--- pass-manager-tkinter/launcher.py
@@ -1,3 +1,0 @@
-from scripts import ui
-
-ui_obj = ui.UI()
