--- pass-manager-tkinter/requirements.txt
@@ -1,9 +1,0 @@
-cffi==1.14.5
-cryptography==3.4.6
-numpy==1.20.1
-pandas==1.2.3
-pycparser==2.20
-pyperclip==1.8.2
-python-dateutil==2.8.1
-pytz==2021.1
-six==1.15.0
